test: move post-publish checklist tests into a dedicated class
The checklist tests were bolted onto Write_Test.php (which tests write.php),
while the sibling email-verification feature got its own parallel test class.
Extract them into Write_Post_Publish_Checklist_Test.php so the test layout
mirrors the src layout (one class per source file) and matches the
email-verification precedent. Moves the 11 tests, the go_to_singular_post()
helper, and the checklist-specific tear_down state out of Write_Test.php; the
new class require_once's write.php explicitly (its helpers + the code under
test). No assertion changes.
